expose writer interface
