+++ cynic-parser/src/lib.rs
@@ -4,13 +4,16 @@ mod ast;
 mod lexer;
 mod printer;
 
-pub use self::ast::{ids, readers, Ast, AstBuilder};
+// TODO: Rethink WrappingType, OperationType & DirectiveLocation at least?
+pub use self::ast::{
+    ids, readers, storage, writer::AstWriter, Ast, DirectiveLocation, OperationType, WrappingType,
+};
 
 lalrpop_mod!(schema);
 
 pub fn parse_type_system_document(input: &str) -> Ast {
     let lexer = lexer::Lexer::new(input);
-    let mut ast = AstBuilder::new();
+    let mut ast = AstWriter::new();
 
     schema::TypeSystemDocumentParser::new()
         .parse(input, &mut ast, lexer)
